feat: embed company contact form at bottom of /bedrift page (#169)
Replaces the CTA that linked out to a separate /bedrift/kontakt page
with the actual contact form, so companies can send an inquiry
without leaving the landing page. The dedicated route is removed;
in-page links now scroll to the form via a #kontakt anchor.
